Remove showError case in getToken
Remove error in getToken for props that describe a native currency.
Only show an error if there is a mismatch between desired token and the specified chain/wallet.
Consolidate duplicate methods into CurrencyInfoHelpers.
